requirements.txt
wxpython
pypubsub
